% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbboxplot.R
\name{svydbboxplot}
\alias{svydbboxplot}
\title{Boxplot}
\usage{
svydbboxplot(x, groups = NULL, design, varwidth = F, outlier = F,
  all.outlier = F)
}
\arguments{
\item{x}{Name indicating the variable.}

\item{groups}{Ifgroupsis defined, boxes of x will be split bygroups.}

\item{design}{\code{svydb.design} object.}

\item{varwidth}{if \code{varwidth = T},  the  width of the boxes are proportional
to the number of observations in that box.}

\item{outlier}{if \code{outlier = T}, observations outside of the 1.5 * IQR will be
plotted as points.}

\item{all.outlier}{if \code{all.outlier = F}, Not all the outliers will be plotted.
#' @description
#' Boxplots for survey data sets.}
}
\description{
Boxplot
}
\examples{
data(nhane)
nh.dbsurv = svydbdesign(st = SDMVSTRA, wt = WTMEC2YR, id = SDMVPSU, data = nhane)
svydbboxplot(x = Age, groups = Gender, design = nh.dbsurv)
svydbboxplot(x = Age, groups = Race3, design = nh.dbsurv, varwidth = T)
svydbboxplot(x = Height, groups = Race3, design = nh.dbsurv, outlier = T)
svydbboxplot(x = Weight, groups = Gender, design = nh.dbsurv,
     outlier = T, all.outlier = T, varwidth = T)
# OR with a database connection
# library(MonetDBLite)
# library(DBI)
# library(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "nhane", nhane)
# nhane.db = tbl(con, "nhane")
# nh.dbsurv = svydbdesign(st = SDMVSTRA, wt = WTMEC2YR, id = SDMVPSU, data = nhane.db)
}
\seealso{
\code{\link{svydbdesign}}, \code{\link{svydbquantile}}
}
\author{
Charco Hui
}
