% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbby.R
\name{svydbby}
\alias{svydbby}
\title{Survey statistics on subsets}
\usage{
svydbby(x, by, FUN, design, ...)
}
\arguments{
\item{x}{A variable specifying the variable to pass to FUN.}

\item{by}{A variable specifying factors that define the subsets.}

\item{FUN}{A function indicating the desired survey statistics.}

\item{design}{\code{svydb.design} object.}

\item{...}{Other arguments to pass to FUN.}
}
\description{
Survey statistics on subsets, currently only works on \code{\link{svydbtotal}},
\code{\link{svydbmean}}.
}
\examples{
data(nhane)
nh.dbsurv = svydbdesign(st = SDMVSTRA , wt = WTMEC2YR,id = SDMVPSU , data = nhane)
svydbby(x = Age , by = Gender , FUN = svydbmean ,design = nh.dbsurv , num = T)
svydbby(x = BMI , by = Race3, FUN = svydbtotal ,design = nh.dbsurv , num = T)
# OR with a database connection
# require(MonetDBLite)
# require(DBI)
# require(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "nhane", nhane)
# nhane.db = tbl(con, "nhane")
# nh.dbsurv = svydbdesign(st = SDMVSTRA , wt = WTMEC2YR,id = SDMVPSU , data = nhane.db)
}
\seealso{
\code{\link{svydbdesign}}, \code{\link{svydbtotal}}, \code{\link{svydbmean}}
}
\author{
Charco Hui
}
