% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbcoplot.R
\name{svydbcoplot}
\alias{svydbcoplot}
\title{Conditional plots}
\usage{
svydbcoplot(formula, by, design)
}
\arguments{
\item{formula}{Formula indicating x and y. i.e. y~x.}

\item{by}{Formula indicating the conditions of each plot. i.e by~by2}

\item{design}{\code{svydb.design} object.}
}
\description{
Hexagon binning will be used for conditional plots.
}
\examples{
data(nhane)
nh.dbsurv = svydbdesign(st = SDMVSTRA , wt = WTMEC2YR,id = SDMVPSU , data = nhane)
svydbcoplot(Age~Height , by = SmokeNow~Gender ,design = nh.dbsurv)
# OR with a database connection
# library(MonetDBLite)
# library(DBI)
# library(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "nhane", nhane)
# nhane.db = tbl(con, "nhane")
# nh.dbsurv = svydbdesign(st = SDMVSTRA , wt = WTMEC2YR,id = SDMVPSU , data = nhane.db)
}
\seealso{
\code{\link{svydbdesign}}, \code{\link{svydbhexbin}}, \code{\link{svydbhexplot}}
}
\author{
Charco Hui
}
