% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbdesign.R
\name{svydbdesign}
\alias{svydbdesign}
\title{Survey design}
\usage{
svydbdesign(st = NULL, id = NULL, wt = NULL, data)
}
\arguments{
\item{st}{Column name specifying the strata column. \code{NULL} for no strata.}

\item{id}{Column name specifying the cluster column. \code{NULL} for no cluster.}

\item{wt}{Column name specifying the sampling weights column.}

\item{data}{A data frame or sql table of the survey data set..}
}
\description{
Gathers all information that are needed to compute survey statistics
into a design.
}
\examples{
data(nhane)
nh.dbsurv = svydbdesign(st = SDMVSTRA , wt = WTMEC2YR,id = SDMVPSU , data = nhane)
nh.dbsurv$subset(Race3 == 3)
nh.dbsurv$getmh()
nh.dbsurv$getwt()
nh.dbsurv$clone()
# OR with a database connection
# require(MonetDBLite)
# require(DBI)
# require(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "nhane", nhane)
# nhane.db = tbl(con, "nhane")
}
\seealso{
\code{\link{svydbtotal}}, \code{\link{svydbmean}}
}
\author{
Charco Hui
}
