% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbhexbin.R
\name{svydbhexplot}
\alias{svydbhexplot}
\title{Plotting hexagons.}
\usage{
svydbhexplot(d, xlab = d$xlab, ylab = d$ylab)
}
\arguments{
\item{d}{returning object of \code{svydbhexbin}.}

\item{xlab, ylab}{labels for x and y axis.}
}
\description{
Ploting the hexagons where the density within each hexagon is proportional to their sampling weights.
}
\examples{
data(nhane)
nh.dbsurv = svydbdesign(st = SDMVSTRA, wt = WTMEC2YR, id = SDMVPSU, data = nhane)
hb = svydbhexbin(Height ~ Weight, design = nh.dbsurv)
svydbhexplot(hb, xlab = "xx", ylab = "yy")
svydbhexplot(hb) + theme_bw() + ggtitle("Hex")
# OR with a database connection
# library(MonetDBLite)
# library(DBI)
# library(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "nhane", nhane)
# nhane.db = tbl(con, "nhane")
# nh.dbsurv = svydbdesign(st = SDMVSTRA, wt = WTMEC2YR, id = SDMVPSU, data = nhane.db)
}
\seealso{
\code{\link{svydbdesign}}, \code{\link{svydbhexbin}}
}
\author{
Charco Hui
}
