% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbhist.R
\name{svydbhist}
\alias{svydbhist}
\title{Histogram}
\usage{
svydbhist(x, design, binwidth = NULL, xlab = "x", ylab = "Density",
  ...)
}
\arguments{
\item{x}{Name indicating the variable.}

\item{design}{\code{svydb.design} object.}

\item{binwidth}{The width of each  bin. Binswidths are calculated with Sturges’ formula.}

\item{xlab, ylab}{labels for xlab and ylab.}
}
\description{
Histograms weighted by sampling weights.
}
\examples{
data(nhane)
nh.dbsurv = svydbdesign(st = SDMVSTRA, wt = WTMEC2YR, id = SDMVPSU, data = nhane)
svydbhist(x = DirectChol, design = nh.dbsurv)
svydbhist(x = DirectChol, design = nh.dbsurv, binwidth = 0.25)
svydbhist(x = Height, design = nh.dbsurv, xlab = "xx", ylab = "yy")
svydbhist(x = Weight, design = nh.dbsurv) + ggtitle("Weight")
# OR with a database connection
# library(MonetDBLite)
# library(DBI)
# library(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "nhane", nhane)
# nhane.db = tbl(con, "nhane")
# nh.dbsurv = svydbdesign(st = SDMVSTRA, wt = WTMEC2YR, id = SDMVPSU, data = nhane.db)
}
\seealso{
\code{\link{svydbdesign}}, \code{\link{svydbmean}}
}
\author{
Charco Hui
}
