% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbmean.R
\name{svydbmean}
\alias{svydbmean}
\title{Population Mean}
\usage{
svydbmean(x, num, design, return.mean = F,
  lonely.psu = getOption("svydb.lonely.psu"), ...)
}
\arguments{
\item{x}{Name indicating the variable.}

\item{num}{\code{TRUE} or \code{FALSE} indicating whether x is numeric or categorical.}

\item{design}{\code{svydb.design} object.}

\item{return.mean}{\code{TRUE} to return only means, no standard errors.}
}
\description{
Computes population mean from survey data sets.
}
\examples{
data(nhane)
nh.dbsurv = svydbdesign(st = SDMVSTRA , wt = WTMEC2YR,id = SDMVPSU , data = nhane)
svydbmean(x = DirectChol, design = nh.dbsurv , num = T)
svydbmean(x = Race3, design = nh.dbsurv , num = F)
svydbmean(x = DirectChol, design = nh.dbsurv , num = T, return.mean = T)
coef(svydbmean(x = DirectChol, design = nh.dbsurv , num = T))
SE(svydbmean(x = DirectChol, design = nh.dbsurv , num = T))
# OR with a database connection
# library(MonetDBLite)
# library(DBI)
# library(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "nhane", nhane)
# nhane.db = tbl(con, "nhane")
# nh.dbsurv = svydbdesign(st = SDMVSTRA , wt = WTMEC2YR,id = SDMVPSU , data = nhane.db)
}
\seealso{
\code{\link{svydbdesign}}, \code{\link{svydbtotal}}
}
\author{
Charco Hui
}
