% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbquantile.R
\name{svydbquantile}
\alias{svydbquantile}
\title{Quantiles}
\usage{
svydbquantile(x, quantiles = 0.5, design)
}
\arguments{
\item{x}{Name indicating the variable.}

\item{quantiles}{Quantiles to estimate, a number, or a vector of numbers for multiple quantiles. Default to 0.5.}

\item{design}{\code{svydb.design} object.}
}
\description{
Compute quantiles from survey data sets.
}
\examples{
data(nhane)
nh.dbsurv = svydbdesign(st = SDMVSTRA , wt = WTMEC2YR,id = SDMVPSU , data = nhane)
svydbquantile(x = Age , quantile = 0.5, design = nh.dbsurv)
svydbquantile(x = Age , quantile = c(0.25, 0.75), design = nh.dbsurv)
# OR with a database connection
# require(MonetDBLite)
# require(DBI)
# require(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "nhane", nhane)
# nhane.db = tbl(con, "nhane")
# nh.dbsurv = svydbdesign(st = SDMVSTRA , wt = WTMEC2YR,id = SDMVPSU , data = nhane.db)
}
\seealso{
\code{\link{svydbdesign}}
}
\author{
Charco Hui
}
