% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbrepdesign.R
\name{svydbrepdesign}
\alias{svydbrepdesign}
\title{Survey replicate design}
\usage{
svydbrepdesign(st = NULL, id = NULL, wt = NULL, repwt = NULL, scale,
  data)
}
\arguments{
\item{st}{Column name specifying the strata column. \code{NULL} for no strata.}

\item{id}{Column name specifying the cluster column. \code{NULL} for no cluster.}

\item{wt}{Column name specifying the sampling weights column.}

\item{repwt}{Regular expressions that matches the columns of the replicate weights.}

\item{scale}{Scaling constant for variance.}

\item{data}{A data frame or sql table of the survey data set.}
}
\description{
Gathers all information that are needed to compute survey statistics
into a design. Currently, only \code{\link{svydbreptotal}}, \code{\link{svydbrepmean}},
are available for replicate statistics.
}
\examples{
data(ss16hde)
hde.dbrepsurv = svydbrepdesign(wt = WGTP, repwt="wgtp[0-9]+", scale = 4/80, data = ss16hde)
hde.dbrepsurv$subset(BATH == 1)
hde.dbrepsurv$clone()
# OR with a database connection
# require(MonetDBLite)
# require(DBI)
# require(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "ss16hde", ss16hde)
# ss16hde.db = tbl(con, "ss16hde")
# hde.dbrepsurv = svydbrepdesign(wt = WGTP, repwt="wgtp[0-9]+", scale = 4/80, data = ss16hde.db)
}
\author{
Charco Hui
}
