% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbrepmean.R
\name{svydbrepmean}
\alias{svydbrepmean}
\title{Population Mean with replicate weights}
\usage{
svydbrepmean(x, design, num, return.replicates = F)
}
\arguments{
\item{x}{Name indicating the variable.}

\item{design}{\code{svydb.repdesign} object.}

\item{num}{\code{TRUE} or \code{FALSE} indicating whether x is numeric or categorical.}

\item{return.replicates}{\code{TRUE} to return the means computed with replicate weights.}
}
\description{
Computes population mean from survey data sets with replicate weights.
}
\examples{
data(ss16hde)
hde.dbrepsurv = svydbrepdesign(wt = WGTP, repwt="wgtp[0-9]+", scale = 4/80, data = ss16hde)
svydbrepmean(x = BATH, design = hde.dbrepsurv, num = T)
svydbrepmean(x = FS, design = hde.dbrepsurv, num = F)
svydbrepmean(x = HHT, design = hde.dbrepsurv, num = T, return.replicates = T)$replicates
coef(svydbrepmean(x = BATH, design = hde.dbrepsurv, num = T))
SE(svydbrepmean(x = BATH, design = hde.dbrepsurv, num = T))
# OR with a database connection
# library(MonetDBLite)
# library(DBI)
# library(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "ss16hde", ss16hde)
# ss16hde.db = tbl(con, "ss16hde")
# hde.dbrepsurv = svydbrepdesign(wt = WGTP, repwt="wgtp[0-9]+", scale = 4/80, data = ss16hde.db)
}
\seealso{
\code{\link{svydbrepdesign}}, \code{\link{svydbreptotal}}
}
\author{
Charco Hui
}
