% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbreptotal.R
\name{svydbreptotal}
\alias{svydbreptotal}
\title{Population Total with replicate weights}
\usage{
svydbreptotal(x, design, num, return.replicates = F)
}
\arguments{
\item{x}{Name indicating the variable.}

\item{design}{\code{svydb.repdesign} object.}

\item{num}{\code{TRUE} or \code{FALSE} indicating whether x is numeric or categorical.}

\item{return.replicates}{\code{TRUE} to return the totals computed with replicate weights.}
}
\description{
Computes population total from survey data sets with replicate weights.
}
\examples{
data(ss16hde)
hde.dbrepsurv = svydbrepdesign(wt = WGTP, repwt="wgtp[0-9]+", scale = 4/80, data = ss16hde)
svydbreptotal(x = BATH, design = hde.dbrepsurv , num = T)
svydbreptotal(x = FS, design = hde.dbrepsurv , num = F)
svydbreptotal(x = HHT, design = hde.dbrepsurv , num = T, return.replicates = T)$replicates
coef(svydbreptotal(x = BATH, design = hde.dbrepsurv , num = T))
SE(svydbreptotal(x = BATH, design = hde.dbrepsurv , num = T))
# OR with a database connection
# require(MonetDBLite)
# require(DBI)
# require(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "ss16hde", ss16hde)
# ss16hde.db = tbl(con, "ss16hde")
# hde.dbrepsurv = svydbrepdesign(wt = WGTP, repwt="wgtp[0-9]+", scale = 4/80, data = ss16hde.db)
}
\seealso{
\code{\link{svydbrepdesign}}, \code{\link{svydbreptotal}}
}
\author{
Charco Hui
}
