% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbtable.R
\name{svydbtable}
\alias{svydbtable}
\title{Survey Tables}
\usage{
svydbtable(formula, design, as.local = F)
}
\arguments{
\item{formula}{A formula specifying margins for the table, only + can be used.}

\item{design}{\code{svydb.design} object.}

\item{as.local}{A logical value indicating the returning object type.
Default is database tables, \code{tbl_sql} if the input is \code{tbl_sql}.}
}
\description{
Create contingency tables for survey data sets.
}
\examples{
data(nhane)
nh.dbsurv = svydbdesign(st = SDMVSTRA, wt = WTMEC2YR,id = SDMVPSU, data = nhane)
svydbtable(~Race3 + Smoke100 + Gender, nh.dbsurv)
svydbtable(~MaritalStatus, design = nh.dbsurv, as.local = T)
# OR with a database connection
# library(MonetDBLite)
# library(DBI)
# library(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "nhane", nhane)
# nhane.db = tbl(con, "nhane")
# nh.dbsurv = svydbdesign(st = SDMVSTRA, wt = WTMEC2YR, id = SDMVPSU, data = nhane.db)
}
\seealso{
\code{\link{svydbdesign}}
}
\author{
Charco Hui
}
