% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svydbtotal.R
\name{svydbtotal}
\alias{svydbtotal}
\title{Population Total}
\usage{
svydbtotal(x, num, design, return.total = F,
  lonely.psu = getOption("svydb.lonely.psu"), ...)
}
\arguments{
\item{x}{Name indicating the variable.}

\item{num}{\code{TRUE} or \code{FALSE} indicating whether x is numeric or categorical.}

\item{design}{\code{svydb.design} object.}

\item{return.total}{\code{TRUE} to return only totals, no standard errors.}
}
\description{
Computes population total from survey data sets.
}
\examples{
data(nhane)
nh.dbsurv = svydbdesign(st = SDMVSTRA, wt = WTMEC2YR, id = SDMVPSU, data = nhane)
svydbtotal(x = DirectChol, design = nh.dbsurv, num = T)
svydbtotal(x = Race3, design = nh.dbsurv, num = F)
svydbtotal(x = DirectChol, design = nh.dbsurv, num = T, return.total = T)
coef(svydbtotal(x = DirectChol, design = nh.dbsurv, num = T))
SE(svydbtotal(x = DirectChol, design = nh.dbsurv, num = T))
# OR with a database connection
# library(MonetDBLite)
# library(DBI)
# library(dbplyr)
# con = dbConnect(MonetDBLite())
# dbWriteTable(con, "nhane", nhane)
# nhane.db = tbl(con, "nhane")
# nh.dbsurv = svydbdesign(st = SDMVSTRA, wt = WTMEC2YR, id = SDMVPSU, data = nhane.db)
}
\seealso{
\code{\link{svydbdesign}}, \code{\link{svydbmean}}
}
\author{
Charco Hui
}
